api-audio-processing/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-audio-processing"
version = "0.1.0"
description = "API para processamento de áudio (text-to-speech e speech-to-text)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.6",
    "openai>=1.10.0",
    "httpx>=0.25.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.2",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.7.1",
]

[project.scripts]
api-audio-processing = "api_audio_processing.main:main"